small change in e2e test

+++ e2e.sh
@@ -16,7 +16,7 @@ npm i @angular/cli -g
 
 # goto e2e directory
 cd ../../e2e$2
-# echo $PWD
+echo 'tested version is in ' $PWD
 
 # install packages & generate electron project
 npm i ../packages/schematics/electron-schematics-schematics-0.0.1.tgz
@@ -25,12 +25,12 @@ npm i ../packages/build-electron/electron-schematics-build-electron-0.0.1.tgz
 
 # run the app with auto exit
 cp 'main.ts' 'projects/electron/main.ts'
-ng serve electron
+LOG=$(ng serve electron)
 
-# check if log file exists
-# if [[ ! -f e2e.log ]] ; then
-#     exit 1
-# fi
+# check the log
+if [[ ! $LOG == *"Angualr say: on init"* ]] ; then
+    exit 1
+fi
 
 # try to build the app
 rm -r ./dist
@@ -41,4 +41,4 @@ if [[ ! -f $1 ]] ; then
     exit 1
 fi
 
-# read -p "Press enter to continue"
+read -p "Press enter to continue"
